🎨 [Chore] 가온 반올림 기능 추가에 따른 자료형 변경

+++ GoWalk/GoWalk/Detail/Model/DetailWeather.swift
@@ -10,26 +10,26 @@ import UIKit
 struct DetailWeather {
     struct Hourly { // 시간별 날씨
         let time: Int
-        let rawTemperature: Int
+        let rawTemperature: Double
         let iconName: String
 
         var temperature: String {
-            SettingsManager.shared.convertedTemperature(Double(rawTemperature))
+            SettingsManager.shared.convertedTemperature(rawTemperature)
         }
     }
 
     struct Weekly { // 주간 날씨
         let date: Int
         let iconName: String
-        let rawMinTemperature: Int
-        let rawMaxTemperature: Int
+        let rawMinTemperature: Double
+        let rawMaxTemperature: Double
 
         var minTemperature: String {
-            SettingsManager.shared.convertedTemperature(Double(rawMinTemperature))
+            SettingsManager.shared.convertedTemperature(rawMinTemperature)
         }
 
         var maxTemperature: String {
-            SettingsManager.shared.convertedTemperature(Double(rawMaxTemperature))
+            SettingsManager.shared.convertedTemperature(rawMaxTemperature)
         }
     }
 }
